+++ crates/ruma-events/src/room.rs
@@ -23,7 +23,6 @@ pub mod name;
 pub mod pinned_events;
 pub mod power_levels;
 pub mod redaction;
-pub mod relationships;
 pub mod server_acl;
 pub mod third_party_invite;
 pub mod tombstone;